missing cast

+++ src/overlapping_partitioning.cpp
@@ -314,7 +314,7 @@ Clusters OverlappingKMeansPartitioningSPANN(PointSet& points, const Partition& p
         size_t total_num_moves = parlay::reduce(num_moves_into_cluster);
 
         if (total_num_moves > num_assignments_left) {
-            double fraction_to_keep = num_assignments_left / total_num_moves;
+            double fraction_to_keep = static_cast<double>(num_assignments_left) / total_num_moves;
             num_moves_into_cluster = parlay::map(num_moves_into_cluster, [&](size_t num_moves) -> size_t {
                 return std::floor(num_moves * fraction_to_keep);
             });
